+++ docs/conf.py
@@ -38,6 +38,7 @@ with open(os.path.join(base_dir, 'resolwe', '__about__.py')) as f:
 extensions = [
     'sphinx.ext.autodoc',
     'sphinx.ext.doctest',
+    'sphinx.ext.extlinks',
     'sphinx.ext.intersphinx',
     'sphinx.ext.coverage',
     'sphinx.ext.viewcode',
@@ -84,5 +85,14 @@ html_theme = 'sphinx_rtd_theme'
 # Output file base name for HTML help builder.
 htmlhelp_basename = 'Resolwedoc'
 
-# Example configuration for intersphinx: refer to the Python standard library.
-intersphinx_mapping = {'https://docs.python.org/': None}
+# Configuration for intersphinx
+intersphinx_mapping = {
+    'python': ('https://docs.python.org/3', None),
+    'django': ('https://docs.djangoproject.com/en/dev/',
+               'https://docs.djangoproject.com/en/dev/_objects/'),
+}
+
+# Configuration for extlinks
+extlinks = {
+    'drf': ('http://www.django-rest-framework.org/api-guide/%s', ''),
+}