+++ resolwe/flow/management/__init__.py
@@ -1,0 +1,12 @@
+"""
+===============
+Flow Management
+===============
+
+.. automodule:: resolwe.flow.management.commands.purge
+    :members:
+
+.. automodule:: resolwe.flow.management.commands.register
+    :members:
+
+"""