+++ resolwe/flow/management/commands/register.py
@@ -1,4 +1,9 @@
-"""Register processes"""
+"""
+==================
+Register Processes
+==================
+
+"""
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 import jsonschema